refactor(ethereum): split legacy handler module
